// src/app/Shared/model/Address.ts
export interface Address {
    typeId: number;
    street: string;
    city: string;
    countryId: number;
    stateId: number;
    stateOther: string;
    zipCode: string;
    phoneNumberPrimary: string;
    phoneNumberSecondary: string;
    faxNumber: string;
    mobileNumberPrimary: string;
    mobileNumberSecondary: string;
    emaild: string;
    skypeId: string;
    createdBy: number;
    status: string;
    addressLine1: string;
    addressLine2: string;
    country: string;
    state: string;
    countryName:string;
    stateName:string;
}